created new SelectedBeasts file, changed width of image to perfectly fit the modal
